# docs/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: msmariaulfa33@gmail.com
  version: 0.0.1
  title: DNR Service Swagger Docs
  description: Service for DNR Website
servers:
  - description: Local development
    url: 'http://localhost:3001'
...
paths:
...
  /api/v1/products:
    get:
      summary: Get Products with Filter
      operationId: getProducts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  totalPage:
                    type: integer
                  totalProduct:
                    type: integer
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product-summary'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Get array of product by multiple search queries
      parameters:
        - schema:
            type: string
          in: query
          name: category
          description: Get product by category name
        - schema:
            type: string
          in: query
          allowReserved: true
          name: name
          description: Get product by product name
        - schema:
            type: string
          in: query
          name: company_name
          description: Get product by product's company_name
        - schema:
            type: string
          in: query
          name: slug
          description: Get product by product's slug
        - schema:
            type: string
          in: query
          name: page
          description: Get product page use for pagination
        - schema:
            type: string
          in: query
          name: limit
          description: Limit the return product
        - schema:
            type: string
            example: PRICE
            enum:
              - PRICE
              - NAME
              - RATE
              - DISCOUNT
              - TOP_SALES
          in: query
          name: sort_by
          description: Indicating field for sorting the result
        - schema:
            type: string
            enum:
              - ASC
              - DESC
            example: ASC
          in: query
          name: order
          description: Indicating sorting
        - schema:
            type: string
          in: query
          name: min_price
          description: Filtering the minimum price
        - schema:
            type: string
          in: query
          name: max_price
          description: Filtering the maximum price
        - schema:
            type: string
          in: query
          name: rating
          description: Filtering the maximum price
      tags:
        - product
    parameters: []
  /api/v1/products/top-product:
    get:
      summary: Get Top perfoming products
      operationId: getTopProduct
      parameters:
        - schema:
            type: string
          in: query
          name: limit
          description: Limit the flash product return
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product-top'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Get array of product by multiple search queries
      tags:
        - product
    parameters: []
  /api/v1/products/{productId}:
    get:
      summary: Get Product By Id
      operationId: getProductById
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      parameters: []
      description: Get single product by Id
      tags:
        - product
    parameters:
      - schema:
          type: string
        name: productId
        in: path
        required: true
        description: Id of the product
  /api/v1/products/favorites:
    get:
      summary: Get Product Favorites For User
      operationId: getProductFavoritesForUser
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  totalPage:
                    type: integer
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product-summary'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      parameters:
        - schema:
            type: string
          in: query
          name: page
          description: Get product page use for pagination
        - schema:
            type: string
          in: query
          name: limit
          description: Limit the return product
        - schema:
            type: string
            example: PRICE
            enum:
              - PRICE
              - NAME
              - RATE
              - DISCOUNT
          in: query
          name: sort_by
          description: Indicating field for sorting the result
        - schema:
            type: string
            enum:
              - ASC
              - DESC
            example: ASC
          in: query
          name: order
          description: Indicating sorting method
      description: Get list of product favorites
      tags:
        - product
  /api/v1/products/{productId}/favorites:
    post:
      summary: Add a Product to favorites
      operationId: addProductFavorites
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Add product to favorites
      tags:
        - product
    delete:
      summary: Delete product favorites
      operationId: deleteProductFavorites
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Healthcheck'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      parameters: []
      description: Delete product from favorites
      tags:
        - product
    parameters:
      - schema:
          type: string
        name: productId
        in: path
        required: true
        description: Id of the product
  /api/v1/admin/products:
    get:
      summary: Get Products with Filter for Admin
      operationId: getProductsForAdmin
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  totalPage:
                    type: integer
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product-summary'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Get array of product by multiple search queries
      parameters:
        - schema:
            type: string
          in: query
          name: category
          description: Get product by category name
        - schema:
            type: string
          in: query
          name: name
          allowReserved: true
          description: Get product by product name
        - schema:
            type: string
          in: query
          name: company_name
          description: Get product by product's company_name
        - schema:
            type: string
          in: query
          name: slug
          description: Get product by product's slug
        - schema:
            type: string
            enum:
              - ACTIVE
              - INACTIVE
          in: query
          name: status
          description: Get product by product's status
        - schema:
            type: string
          in: query
          name: page
          description: Get product page use for pagination
        - schema:
            type: string
          in: query
          name: limit
          description: Limit the return product
        - schema:
            type: string
            example: PRICE
            enum:
              - PRICE
              - NAME
              - RATE
              - DISCOUNT
          in: query
          name: sort_by
          description: Indicating field for sorting the result
        - schema:
            type: string
            enum:
              - ASC
              - DESC
            example: ASC
          in: query
          name: order
          description: Indicating sorting method
      tags:
        - admin
    post:
      summary: Add a New Product for Admin
      operationId: addProduct
      responses:
        201:
          description: User Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              examples: {}
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Product-create'
            examples: {}
          text/plain:
            schema:
              type: object
              properties: {}
        description: Post the necessary fields for the API to create a new product.
      description: Add a new product
      tags:
        - product
    parameters: []
  /api/v1/admin/products/upload:
    post:
      summary: Create a New Product for Admin
      operationId: uploadProduct
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Create a new category
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      tags:
        - product
  /api/v1/admin/products/{productId}:
    get:
      summary: Get Product By Id for Admin
      operationId: getProductByIdForAdmin
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      parameters: []
      description: Get single product by Id
      tags:
        - product
    patch:
      summary: Update Product For Admin
      operationId: updateProductById
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Update product by Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product-update'
        description: ''
      tags:
        - admin
    parameters:
      - schema:
          type: string
        name: productId
        in: path
        required: true
        description: Id of the product
  /api/v1/admin/products/{product_id}/upload:
    patch:
      tags:
        - upload
      parameters:
        - in: path
          name: product_id
          schema:
            type: string
          required: true
      summary: upload product photo
      operationId: uploadProductPhoto
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Product-summary'
              examples: {}
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        404:
          description: example error 404 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Upload images for product
  /api/v1/admin/products/images/{image_id}:
    parameters:
      - schema:
          type: string
        name: image_id
        in: path
        required: true
        description: id of product images
    delete:
      summary: Delete product image for Admin
      operationId: deleteProductImageForAdmin
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Healthcheck'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      parameters: []
      description: Delete product image by Id
      tags:
        - product
  /api/v1/admin/products/sync-product:
    post:
      summary: Sync product data with SAP
      operationId: syncProduct
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Add a new product
      tags:
        - product
    parameters: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-auth-token
  schemas:
    Healthcheck:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: OK
...
    Error400Response:
      type: object
      required:
        - errorCode
        - message
      properties:
        errorCode:
          type: string
          enum:
            - INVALID_API_KEY
            - API_VALIDATION_ERROR
            - CAN_NOT_REGISTER
            - CAN_NOT_LOGIN
            - UNAUTHORIZED_USER
            - OUTLET_CREATION_ERROR
            - CREATE_ORDER_ERROR
            - ADDRESS_NOT_FOUND
            - ADD_TO_CART_ERROR
            - UPDATE_ORDER_ERROR
            - COMPLETE_PAYMENT_ORDER_ERROR
            - CANCEL_ORDER_ERROR
            - COMPLETE_ORDER_ERROR
            - FLASH_SALE_ERROR
            - EXISTING_CUSTOMER_ID
            - RESET_PASSWORD_ERROR
            - NEWS_ERROR
            - OUTLET_TYPES_ERROR
            - PROMOTION_PRODUCTS_EXIST
            - PROMOTION_CODE_EXIST
        message:
          type: string
        errors:
          $ref: '#/components/schemas/ArrayOfErrors'
...
    Error500Response:
      type: object
      required:
        - errorCode
        - message
      properties:
        errorCode:
          type: string
          enum:
            - INTERNAL_SERVER_ERROR
        message:
          type: string
        errors:
          $ref: '#/components/schemas/ArrayOfErrors'
...
    ArrayOfErrors:
      nullable: true
      type: array
      items:
        oneOf:
          - type: string
          - type: object
...
    Product-summary:
      title: ProductSummary
      type: object
      x-examples:
        CDR:
          value:
            name: CDR (Calcium-D-Redoxon) ahlinya kalsium suplemen makanan
            picture_url: cdr-pic-01
            company_name: CDR
            unit: PC
            slug: cdr-ahlinya-kalsium-suplemen
            price: 50000
      description: Model representing summary information of product for carts & order
      properties:
        id:
          type: string
          description: Id of the product
        name:
          type: string
          description: Product title
        picture_url:
          type: string
          nullable: true
          description: Path to picture file
        slug:
          type: string
        unit:
          type: string
          description: Unit of the product
        price:
          type: integer
          nullable: true
          description: Price per unit of product
        discount_percentage:
          type: integer
          nullable: true
          description: Percentage of the discount
        discount_price:
          type: integer
          nullable: true
          description: Discount amount
        discount_type:
          type: string
          nullable: true
          enum:
            - PRICE
            - PERCENTAGE
            - null
          description: Type of the discount to be shown
      required:
        - name
        - slug
        - unit
...